Fix add_function_array() separation
result may be a slot in op2. In that case SEPARATE_ARRAY() will change both
result and the slot in op2. Looping over op2 and inserting the element results
in both reference-less recursion which we don't allow, and increasing the
refcount to 2, failing any further insertions into the array.

Avoid this by copying result into a temporary zval and performing separation
there instead.

Fixes GH-10963