Fix reflection
Instead of modifying the current call frame, we use the trampoline we already
use for parent calls, and push a new stack frame. Unfortunately, this requires
setting scope of the function, which modifies some error messages. But because
A::get() is a confusing method, we now also need to allocate a new string for
$prop::get(), so that the full name in the message becomes A::$prop::get().