Improve performance of property hooks (#133)
* Hooks without guards

* Skip stack overflowing tests on asan

* Avoid re-entry for non-ref hooks

* VM_ENTER fixes

* Fix more bugs

* Fix guard check for other objects

* Cache simple getter calls

* Fix reflection

Instead of modifying the current call frame, we use the trampoline we already
use for parent calls, and push a new stack frame. Unfortunately, this requires
setting scope of the function, which modifies some error messages. But because
A::get() is a confusing method, we now also need to allocate a new string for
$prop::get(), so that the full name in the message becomes A::$prop::get().

* Fix stack overflow tests on windows

* Avoid trampoline for {get,set}RawValue() for unhooked props

* Simplify duplicate check

* Add fast failure for missing cache_slot

* Tweak ASAN skip message